Reverts version, casts web3Tx.to to <string>, moved ExtendedBridgeTx to it's own file.
