+++ src/services/extended-bridge-tx.ts
@@ -1,0 +1,7 @@
+import {Transaction} from 'bridge-transaction-parser';
+
+export default interface ExtendedBridgeTx extends Transaction {
+  blockHash: string;
+  createdOn: Date;
+  to: string;
+}
