+++ src/services/filtered-bridge-transaction-processor.ts
@@ -1,5 +1,5 @@
 import {BridgeDataFilterModel} from '../models/bridge-data-filter.model';
-import { ExtendedBridgeTx } from './node-bridge-data.provider';
+import ExtendedBridgeTx from './extended-bridge-tx';
 
 export default interface FilteredBridgeTransactionProcessor {
     process(transaction: ExtendedBridgeTx): Promise<void>;
